chore(normal): split file and add comments
